+++ program/src/matching.rs
@@ -2,6 +2,7 @@ use std::cell::RefMut;
 use std::convert::TryFrom;
 use std::mem::size_of;
 
+use anchor_lang::prelude::emit;
 use bytemuck::{cast, cast_mut, cast_ref};
 use fixed::types::I80F48;
 use num_enum::{IntoPrimitive, TryFromPrimitive};
@@ -15,13 +16,17 @@ use solana_program::sysvar::Sysvar;
 use static_assertions::const_assert_eq;
 
 use mango_common::Loadable;
+use mango_logs::{mango_emit, ReferralFeeAccrualLog};
 use mango_macro::{Loadable, Pod};
 
 use crate::error::{check_assert, MangoError, MangoErrorCode, MangoResult, SourceFileId};
+use crate::ids::mngo_token;
 use crate::queue::{EventQueue, FillEvent, OutEvent};
 use crate::state::{
-    DataType, MangoAccount, MetaData, PerpMarket, PerpMarketInfo, MAX_PERP_OPEN_ORDERS,
+    DataType, MangoAccount, MangoCache, MangoGroup, MetaData, PerpMarket, PerpMarketCache,
+    PerpMarketInfo, CENTIBPS_PER_UNIT, MAX_PERP_OPEN_ORDERS, ZERO_I80F48,
 };
+use crate::utils::emit_perp_balances;
 
 declare_check_assert_macros!(SourceFileId::Matching);
 pub type NodeHandle = u32;
@@ -757,9 +762,12 @@ impl<'a> Book<'a> {
     #[inline(never)]
     pub fn new_order(
         &mut self,
+        program_id: &Pubkey,
+        mango_group: &MangoGroup,
+        mango_group_pk: &Pubkey,
+        mango_cache: &MangoCache,
         event_queue: &mut EventQueue,
         market: &mut PerpMarket,
-        info: &PerpMarketInfo,
         oracle_price: I80F48,
         mango_account: &mut MangoAccount,
         mango_account_pk: &Pubkey,
@@ -770,12 +778,16 @@ impl<'a> Book<'a> {
         order_type: OrderType,
         client_order_id: u64,
         now_ts: u64,
+        referrer_mango_account_ai: Option<&AccountInfo>,
     ) -> MangoResult {
         match side {
             Side::Bid => self.new_bid(
+                program_id,
+                mango_group,
+                mango_group_pk,
+                mango_cache,
                 event_queue,
                 market,
-                info,
                 oracle_price,
                 mango_account,
                 mango_account_pk,
@@ -785,11 +797,15 @@ impl<'a> Book<'a> {
                 order_type,
                 client_order_id,
                 now_ts,
+                referrer_mango_account_ai,
             ),
             Side::Ask => self.new_ask(
+                program_id,
+                mango_group,
+                mango_group_pk,
+                mango_cache,
                 event_queue,
                 market,
-                info,
                 oracle_price,
                 mango_account,
                 mango_account_pk,
@@ -799,6 +815,7 @@ impl<'a> Book<'a> {
                 order_type,
                 client_order_id,
                 now_ts,
+                referrer_mango_account_ai,
             ),
         }
     }
@@ -830,10 +847,13 @@ impl<'a> Book<'a> {
                 (true, true, price)
             }
         };
-        let native_price = market.lot_to_native_price(price);
-        if native_price.checked_div(oracle_price).unwrap() > info.maint_liab_weight {
-            msg!("Posting on book disallowed due to price limits");
-            post_allowed = false;
+        if post_allowed {
+            // price limit check computed lazily to save CU on average
+            let native_price = market.lot_to_native_price(price);
+            if native_price.checked_div(oracle_price).unwrap() > info.maint_liab_weight {
+                msg!("Posting on book disallowed due to price limits");
+                post_allowed = false;
+            }
         }
 
         let mut rem_quantity = quantity; // base lots (aka contracts)
@@ -911,10 +931,13 @@ impl<'a> Book<'a> {
                 (true, true, price)
             }
         };
-        let native_price = market.lot_to_native_price(price);
-        if native_price.checked_div(oracle_price).unwrap() < info.maint_asset_weight {
-            msg!("Posting on book disallowed due to price limits");
-            post_allowed = false;
+        if post_allowed {
+            // price limit check computed lazily to save CU on average
+            let native_price = market.lot_to_native_price(price);
+            if native_price.checked_div(oracle_price).unwrap() < info.maint_asset_weight {
+                msg!("Posting on book disallowed due to price limits");
+                post_allowed = false;
+            }
         }
 
         let mut rem_quantity = quantity; // base lots (aka contracts)
@@ -970,9 +993,12 @@ impl<'a> Book<'a> {
     #[inline(never)]
     fn new_bid(
         &mut self,
+        program_id: &Pubkey,
+        mango_group: &MangoGroup,
+        mango_group_pk: &Pubkey,
+        mango_cache: &MangoCache,
         event_queue: &mut EventQueue,
         market: &mut PerpMarket,
-        info: &PerpMarketInfo,
         oracle_price: I80F48,
         mango_account: &mut MangoAccount,
         mango_account_pk: &Pubkey,
@@ -982,6 +1008,7 @@ impl<'a> Book<'a> {
         order_type: OrderType,
         client_order_id: u64,
         now_ts: u64,
+        referrer_mango_account_ai: Option<&AccountInfo>,
     ) -> MangoResult {
         // TODO proper error handling
         // TODO handle the case where we run out of compute (right now just fails)
@@ -999,12 +1026,22 @@ impl<'a> Book<'a> {
                 (true, true, price)
             }
         };
-        let native_price = market.lot_to_native_price(price);
-        if native_price.checked_div(oracle_price).unwrap() > info.maint_liab_weight {
-            msg!("Posting on book disallowed due to price limits");
-            post_allowed = false;
+        let info = &mango_group.perp_markets[market_index];
+        if post_allowed {
+            // price limit check computed lazily to save CU on average
+            let native_price = market.lot_to_native_price(price);
+            if native_price.checked_div(oracle_price).unwrap() > info.maint_liab_weight {
+                msg!("Posting on book disallowed due to price limits");
+                post_allowed = false;
+            }
         }
 
+        // referral fee related variables
+        let mut ref_fee_rate = None;
+        let mut referrer_mango_account_opt = None;
+        let mut total_quote_taken = 0;
+
+        // generate new order id
         let order_id = market.gen_order_id(Side::Bid, price);
 
         // if post only and price >= best_ask, return
@@ -1030,10 +1067,28 @@ impl<'a> Book<'a> {
             let match_quantity = rem_quantity.min(best_ask.quantity);
             rem_quantity -= match_quantity;
             best_ask.quantity -= match_quantity;
+            let match_quote = match_quantity * best_ask_price;
+            total_quote_taken += match_quote;
 
-            mango_account.perp_accounts[market_index]
-                .add_taker_trade(match_quantity, -match_quantity * best_ask_price);
+            mango_account.perp_accounts[market_index].add_taker_trade(match_quantity, -match_quote);
             let maker_out = best_ask.quantity == 0;
+
+            // if ref_fee_rate is none, determine it
+            // if ref_valid, then pay into referrer, else pay to perp market
+            if ref_fee_rate.is_none() {
+                let (a, b) = determine_ref_vars(
+                    program_id,
+                    mango_group,
+                    mango_group_pk,
+                    mango_cache,
+                    mango_account,
+                    referrer_mango_account_ai,
+                    now_ts,
+                )?;
+                ref_fee_rate = Some(a);
+                referrer_mango_account_opt = b;
+            }
+
             let fill = FillEvent::new(
                 Side::Bid,
                 best_ask.owner_slot,
@@ -1049,7 +1104,7 @@ impl<'a> Book<'a> {
                 *mango_account_pk,
                 order_id,
                 client_order_id,
-                info.taker_fee,
+                info.taker_fee + ref_fee_rate.unwrap(),
                 best_ask_price,
                 match_quantity,
                 best_ask.version,
@@ -1113,26 +1168,46 @@ impl<'a> Book<'a> {
             mango_account.add_order(market_index, Side::Bid, &new_bid)?;
         }
 
+        // if there were matched taker quote apply ref fees
+        // we know ref_fee_rate is not None if total_quote_taken > 0
+        if total_quote_taken > 0 {
+            apply_taker_fees(
+                market,
+                info,
+                mango_account,
+                mango_account_pk,
+                market_index,
+                referrer_mango_account_opt,
+                referrer_mango_account_ai,
+                total_quote_taken,
+                ref_fee_rate.unwrap(),
+                &mango_cache.perp_market_cache[market_index],
+            );
+        }
+
         Ok(())
     }
 
     #[inline(never)]
     pub fn new_ask(
         &mut self,
+        program_id: &Pubkey,
+        mango_group: &MangoGroup,
+        mango_group_pk: &Pubkey,
+        mango_cache: &MangoCache,
         event_queue: &mut EventQueue,
         market: &mut PerpMarket,
-        info: &PerpMarketInfo,
         oracle_price: I80F48,
         mango_account: &mut MangoAccount,
         mango_account_pk: &Pubkey,
         market_index: usize,
         price: i64,
-        quantity: i64, // quantity is guaranteed to be greater than zero due to initial check --
+        quantity: i64, // quantity is guaranteed to be greater than zero due to initial check
         order_type: OrderType,
         client_order_id: u64,
         now_ts: u64,
+        referrer_mango_account_ai: Option<&AccountInfo>,
     ) -> MangoResult {
-        // TODO proper error handling
         let (post_only, mut post_allowed, price) = match order_type {
             OrderType::Limit => (false, true, price),
             OrderType::ImmediateOrCancel => (false, false, price),
@@ -1147,12 +1222,22 @@ impl<'a> Book<'a> {
                 (true, true, price)
             }
         };
-        let native_price = market.lot_to_native_price(price);
-        if native_price.checked_div(oracle_price).unwrap() < info.maint_asset_weight {
-            msg!("Posting on book disallowed due to price limits");
-            post_allowed = false;
+        let info = &mango_group.perp_markets[market_index];
+        if post_allowed {
+            // price limit check computed lazily to save CU on average
+            let native_price = market.lot_to_native_price(price);
+            if native_price.checked_div(oracle_price).unwrap() < info.maint_asset_weight {
+                msg!("Posting on book disallowed due to price limits");
+                post_allowed = false;
+            }
         }
 
+        // referral fee related variables
+        let mut ref_fee_rate = None;
+        let mut referrer_mango_account_opt = None;
+        let mut total_quote_taken = 0;
+
+        // generate new order id
         let order_id = market.gen_order_id(Side::Ask, price);
 
         // if post only and price >= best_ask, return
@@ -1177,9 +1262,27 @@ impl<'a> Book<'a> {
             let match_quantity = rem_quantity.min(best_bid.quantity);
             rem_quantity -= match_quantity;
             best_bid.quantity -= match_quantity;
-            mango_account.perp_accounts[market_index]
-                .add_taker_trade(-match_quantity, match_quantity * best_bid_price);
+
+            let match_quote = match_quantity * best_bid_price;
+            total_quote_taken += match_quote;
+            mango_account.perp_accounts[market_index].add_taker_trade(-match_quantity, match_quote);
             let maker_out = best_bid.quantity == 0;
+
+            // if ref_fee_rate is none, determine it
+            // if ref_valid, then pay into referrer, else pay to perp market
+            if ref_fee_rate.is_none() {
+                let (a, b) = determine_ref_vars(
+                    program_id,
+                    mango_group,
+                    mango_group_pk,
+                    mango_cache,
+                    mango_account,
+                    referrer_mango_account_ai,
+                    now_ts,
+                )?;
+                ref_fee_rate = Some(a);
+                referrer_mango_account_opt = b;
+            }
 
             let fill = FillEvent::new(
                 Side::Ask,
@@ -1196,7 +1299,7 @@ impl<'a> Book<'a> {
                 *mango_account_pk,
                 order_id,
                 client_order_id,
-                info.taker_fee,
+                info.taker_fee + ref_fee_rate.unwrap(),
                 best_bid_price,
                 match_quantity,
                 best_bid.version,
@@ -1259,6 +1362,23 @@ impl<'a> Book<'a> {
 
             let _result = self.asks.insert_leaf(&new_ask)?;
             mango_account.add_order(market_index, Side::Ask, &new_ask)?;
+        }
+
+        // if there were matched taker quote apply ref fees
+        // we know ref_fee_rate is not None if total_quote_taken > 0
+        if total_quote_taken > 0 {
+            apply_taker_fees(
+                market,
+                info,
+                mango_account,
+                mango_account_pk,
+                market_index,
+                referrer_mango_account_opt,
+                referrer_mango_account_ai,
+                total_quote_taken,
+                ref_fee_rate.unwrap(),
+                &mango_cache.perp_market_cache[market_index],
+            );
         }
 
         Ok(())
@@ -1609,4 +1729,107 @@ impl<'a> Book<'a> {
         }
         Ok(())
     }
+}
+
+fn determine_ref_vars<'a>(
+    program_id: &Pubkey,
+    mango_group: &MangoGroup,
+    mango_group_pk: &Pubkey,
+    mango_cache: &MangoCache,
+    mango_account: &MangoAccount,
+    referrer_mango_account_ai: Option<&'a AccountInfo>,
+    now_ts: u64,
+) -> MangoResult<(I80F48, Option<RefMut<'a, MangoAccount>>)> {
+    let mngo_index = match mango_group.find_token_index(&mngo_token::id()) {
+        None => return Ok((ZERO_I80F48, None)),
+        Some(i) => i,
+    };
+
+    let mngo_cache = &mango_cache.root_bank_cache[mngo_index];
+
+    // If the user's MNGO deposit is non-zero then the rootbank cache will be checked already in `place_perp_order`.
+    // If it's zero then cache may be out of date, but it doesn't matter because 0 * index = 0
+    let mngo_deposits = mango_account.get_native_deposit(mngo_cache, mngo_index)?;
+    let ref_mngo_req = I80F48::from_num(mango_group.ref_mngo_required);
+    if mngo_deposits >= ref_mngo_req {
+        return Ok((ZERO_I80F48, None));
+    } else if let Some(referrer_mango_account_ai) = referrer_mango_account_ai {
+        // If referrer_mango_account is invalid, just treat it as if it doesn't exist
+        if let Ok(referrer_mango_account) =
+            MangoAccount::load_mut_checked(referrer_mango_account_ai, program_id, mango_group_pk)
+        {
+            // Need to check if it's valid because user may not have mngo in active assets
+            mngo_cache.check_valid(mango_group, now_ts)?;
+            let ref_mngo_deposits =
+                referrer_mango_account.get_native_deposit(mngo_cache, mngo_index)?;
+
+            if !referrer_mango_account.is_bankrupt
+                && !referrer_mango_account.being_liquidated
+                && ref_mngo_deposits >= ref_mngo_req
+            {
+                return Ok((
+                    I80F48::from_num(mango_group.ref_share_centibps) / CENTIBPS_PER_UNIT,
+                    Some(referrer_mango_account),
+                ));
+            }
+        }
+    }
+    Ok((I80F48::from_num(mango_group.ref_surcharge_centibps) / CENTIBPS_PER_UNIT, None))
+}
+
+fn apply_taker_fees(
+    market: &mut PerpMarket,
+    info: &PerpMarketInfo,
+    mango_account: &mut MangoAccount,
+    mango_account_pk: &Pubkey,
+    market_index: usize,
+    referrer_mango_account_opt: Option<RefMut<MangoAccount>>,
+    referrer_mango_account_ai: Option<&AccountInfo>,
+    total_quote_taken: i64,
+    ref_fee_rate: I80F48,
+    perp_market_cache: &PerpMarketCache,
+) {
+    let taker_quote_native =
+        I80F48::from_num(market.quote_lot_size.checked_mul(total_quote_taken).unwrap());
+
+    if ref_fee_rate > ZERO_I80F48 {
+        let ref_fees = taker_quote_native * ref_fee_rate;
+
+        // if ref mango account is some, then we send some fees over
+        if let Some(mut referrer_mango_account) = referrer_mango_account_opt {
+            mango_account.perp_accounts[market_index].transfer_quote_position(
+                &mut referrer_mango_account.perp_accounts[market_index],
+                ref_fees,
+            );
+            emit_perp_balances(
+                referrer_mango_account.mango_group,
+                *referrer_mango_account_ai.unwrap().key,
+                market_index as u64,
+                &referrer_mango_account.perp_accounts[market_index],
+                perp_market_cache,
+            );
+            mango_emit!(ReferralFeeAccrualLog {
+                mango_group: referrer_mango_account.mango_group,
+                referrer_mango_account: *referrer_mango_account_ai.unwrap().key,
+                referree_mango_account: *mango_account_pk,
+                market_index: market_index as u64,
+                referral_fee_accrual: ref_fees.to_bits(),
+            });
+        } else {
+            // else user didn't have valid amount of MNGO and no valid referrer
+            mango_account.perp_accounts[market_index].quote_position -= ref_fees;
+            market.fees_accrued += ref_fees;
+        }
+    }
+    let taker_fees = taker_quote_native * info.taker_fee;
+    mango_account.perp_accounts[market_index].quote_position -= taker_fees;
+    market.fees_accrued += taker_fees;
+
+    emit_perp_balances(
+        mango_account.mango_group,
+        *mango_account_pk,
+        market_index as u64,
+        &mango_account.perp_accounts[market_index],
+        perp_market_cache,
+    )
 }