+++ program/src/processor.rs
@@ -27,16 +27,17 @@ use mango_common::Loadable;
 use mango_logs::{
     mango_emit, CachePerpMarketsLog, CachePricesLog, CacheRootBanksLog, CancelAllPerpOrdersLog,
     DepositLog, LiquidatePerpMarketLog, LiquidateTokenAndPerpLog, LiquidateTokenAndTokenLog,
-    MngoAccrualLog, OpenOrdersBalanceLog, PerpBalanceLog, PerpBankruptcyLog, RedeemMngoLog,
-    SettleFeesLog, SettlePnlLog, TokenBalanceLog, TokenBankruptcyLog, UpdateFundingLog,
-    UpdateRootBankLog, WithdrawLog,
+    MngoAccrualLog, OpenOrdersBalanceLog, PerpBankruptcyLog, RedeemMngoLog, SettleFeesLog,
+    SettlePnlLog, TokenBalanceLog, TokenBankruptcyLog, UpdateFundingLog, UpdateRootBankLog,
+    WithdrawLog,
 };
 
 use crate::error::{check_assert, MangoError, MangoErrorCode, MangoResult, SourceFileId};
-use crate::ids::msrm_token;
-use crate::ids::srm_token;
+use crate::ids::{msrm_token, srm_token};
 use crate::instruction::MangoInstruction;
 use crate::matching::{Book, BookSide, OrderType, Side};
+#[cfg(not(feature = "devnet"))]
+use crate::oracle::PriceStatus;
 use crate::oracle::{determine_oracle_type, OracleType, Price, StubOracle};
 use crate::queue::{EventQueue, EventType, FillEvent, LiquidateEvent, OutEvent};
 #[cfg(not(feature = "devnet"))]
@@ -44,13 +45,14 @@ use crate::state::PYTH_CONF_FILTER;
 use crate::state::{
     check_open_orders, load_asks_mut, load_bids_mut, load_market_state, load_open_orders,
     load_open_orders_accounts, AdvancedOrderType, AdvancedOrders, AssetType, DataType, HealthCache,
-    HealthType, MangoAccount, MangoCache, MangoGroup, MetaData, NodeBank, PerpAccount, PerpMarket,
-    PerpMarketCache, PerpMarketInfo, PerpTriggerOrder, PriceCache, RootBank, RootBankCache,
-    SpotMarketInfo, TokenInfo, TriggerCondition, UserActiveAssets, ADVANCED_ORDER_FEE,
-    FREE_ORDER_SLOT, INFO_LEN, MAX_ADVANCED_ORDERS, MAX_NODE_BANKS, MAX_PAIRS,
-    MAX_PERP_OPEN_ORDERS, MAX_TOKENS, NEG_ONE_I80F48, ONE_I80F48, QUOTE_INDEX, ZERO_I80F48,
+    HealthType, MangoAccount, MangoCache, MangoGroup, MetaData, NodeBank, PerpMarket,
+    PerpMarketCache, PerpMarketInfo, PerpTriggerOrder, PriceCache, ReferrerIdRecord,
+    ReferrerMemory, RootBank, RootBankCache, SpotMarketInfo, TokenInfo, TriggerCondition,
+    UserActiveAssets, ADVANCED_ORDER_FEE, FREE_ORDER_SLOT, INFO_LEN, MAX_ADVANCED_ORDERS,
+    MAX_NODE_BANKS, MAX_PAIRS, MAX_PERP_OPEN_ORDERS, MAX_TOKENS, NEG_ONE_I80F48, ONE_I80F48,
+    QUOTE_INDEX, ZERO_I80F48,
 };
-use crate::utils::{gen_signer_key, gen_signer_seeds};
+use crate::utils::{emit_perp_balances, gen_signer_key, gen_signer_seeds};
 
 declare_check_assert_macros!(SourceFileId::Processor);
 
@@ -574,7 +576,7 @@ impl Processor {
     }
 
     #[inline(never)]
-    /// *** DEPRECATED Initialize perp market including orderbooks and queues
+    /// DEPRECATED Initialize perp market including orderbooks and queues
     fn add_perp_market(
         program_id: &Pubkey,
         accounts: &[AccountInfo],
@@ -2367,13 +2369,13 @@ impl Processor {
         client_order_id: u64,
         order_type: OrderType,
         reduce_only: bool,
-    ) -> MangoResult<()> {
+    ) -> MangoResult {
         check!(price > 0, MangoErrorCode::InvalidParam)?;
         check!(quantity > 0, MangoErrorCode::InvalidParam)?;
 
         const NUM_FIXED: usize = 8;
-        let accounts = array_ref![accounts, 0, NUM_FIXED + MAX_PAIRS];
-        let (fixed_ais, open_orders_ais) = array_refs![accounts, NUM_FIXED, MAX_PAIRS];
+        let (fixed_ais, open_orders_ais, opt_ais) =
+            array_refs![accounts, NUM_FIXED, MAX_PAIRS; ..;];
         let [
             mango_group_ai,     // read
             mango_account_ai,   // write
@@ -2384,6 +2386,9 @@ impl Processor {
             asks_ai,            // write
             event_queue_ai,     // write
         ] = fixed_ais;
+
+        let referrer_mango_account_ai = opt_ais.first();
+
         let mango_group = MangoGroup::load_checked(mango_group_ai, program_id)?;
 
         let mut mango_account =
@@ -2456,9 +2461,12 @@ impl Processor {
         }
 
         book.new_order(
+            program_id,
+            &mango_group,
+            mango_group_ai.key,
+            &mango_cache,
             &mut event_queue,
             &mut perp_market,
-            &mango_group.perp_markets[market_index],
             mango_cache.get_price(market_index),
             &mut mango_account,
             mango_account_ai.key,
@@ -2469,6 +2477,7 @@ impl Processor {
             order_type,
             client_order_id,
             now_ts,
+            referrer_mango_account_ai,
         )?;
 
         health_cache.update_perp_val(&mango_group, &mango_cache, &mango_account, market_index)?;
@@ -4412,7 +4421,7 @@ impl Processor {
         limit: usize,
     ) -> MangoResult<()> {
         // Limit may be max 10 because of compute limits from logging. Increase if compute goes up
-        let limit = min(limit, 10);
+        let limit = min(limit, 6);
 
         const NUM_FIXED: usize = 4;
         let (fixed_ais, mango_account_ais) = array_refs![accounts, NUM_FIXED; ..;];
@@ -4435,8 +4444,6 @@ impl Processor {
         let perp_market_cache = &mango_cache.perp_market_cache[market_index];
 
         perp_market_cache.check_valid(&mango_group, now_ts)?;
-
-        let info = &mango_group.perp_markets[market_index];
 
         for _ in 0..limit {
             let event = match event_queue.peek_front() {
@@ -4463,20 +4470,8 @@ impl Processor {
                             )?,
                         };
                         let pre_mngo = ma.perp_accounts[market_index].mngo_accrued;
-                        ma.execute_maker(
-                            market_index,
-                            &mut perp_market,
-                            info,
-                            perp_market_cache,
-                            fill,
-                        )?;
-                        ma.execute_taker(
-                            market_index,
-                            &mut perp_market,
-                            info,
-                            perp_market_cache,
-                            fill,
-                        )?;
+                        ma.execute_maker(market_index, &mut perp_market, perp_market_cache, fill)?;
+                        ma.execute_taker(market_index, &mut perp_market, perp_market_cache, fill)?;
                         mango_emit!(MngoAccrualLog {
                             mango_group: *mango_group_ai.key,
                             mango_account: fill.maker,
@@ -4520,14 +4515,12 @@ impl Processor {
                         maker.execute_maker(
                             market_index,
                             &mut perp_market,
-                            info,
                             perp_market_cache,
                             fill,
                         )?;
                         taker.execute_taker(
                             market_index,
                             &mut perp_market,
-                            info,
                             perp_market_cache,
                             fill,
                         )?;
@@ -5277,9 +5270,12 @@ impl Processor {
                 };
 
                 book.new_order(
+                    program_id,
+                    &mango_group,
+                    mango_group_ai.key,
+                    &mango_cache,
                     &mut event_queue,
                     &mut perp_market,
-                    &mango_group.perp_markets[market_index],
                     mango_cache.get_price(market_index),
                     &mut mango_account,
                     mango_account_ai.key,
@@ -5290,6 +5286,7 @@ impl Processor {
                     order.order_type,
                     order.client_order_id,
                     now_ts,
+                    None,
                 )?;
 
                 // TODO OPT - unnecessary, remove after testing
@@ -5319,7 +5316,7 @@ impl Processor {
         program_transfer_lamports(advanced_orders_ai, agent_ai, ADVANCED_ORDER_FEE)
     }
 
-    /// *** Create a MangoAccount PDA and initialize it
+    /// Create a MangoAccount PDA and initialize it
     #[inline(never)]
     fn create_mango_account(
         program_id: &Pubkey,
@@ -5437,7 +5434,7 @@ impl Processor {
         Ok(())
     }
 
-    /// *** Create a DustAccount PDA and initialize it
+    /// Create a DustAccount PDA and initialize it
     #[inline(never)]
     fn create_dust_account(program_id: &Pubkey, accounts: &[AccountInfo]) -> MangoResult {
         const NUM_FIXED: usize = 4;
@@ -5673,6 +5670,147 @@ impl Processor {
         Ok(())
     }
 
+    /// Set the `ref_surcharge_centibps`, `ref_share_centibps` and `ref_mngo_required` on `MangoGroup`
+    #[inline(never)]
+    fn change_referral_fee_params(
+        program_id: &Pubkey,
+        accounts: &[AccountInfo],
+        ref_surcharge_centibps: u32,
+        ref_share_centibps: u32,
+        ref_mngo_required: u64,
+    ) -> MangoResult {
+        check!(ref_surcharge_centibps >= ref_share_centibps, MangoErrorCode::InvalidParam)?;
+
+        const NUM_FIXED: usize = 2;
+        let accounts = array_ref![accounts, 0, NUM_FIXED];
+
+        let [
+            mango_group_ai, // write
+            admin_ai        // read, signer
+        ] = accounts;
+
+        let mut mango_group = MangoGroup::load_mut_checked(mango_group_ai, program_id)?;
+        check_eq!(admin_ai.key, &mango_group.admin, MangoErrorCode::InvalidAdminKey)?;
+        check!(admin_ai.is_signer, MangoErrorCode::SignerNecessary)?;
+        msg!("old referral fee params: ref_surcharge_centibps: {} ref_share_centibps: {} ref_mngo_required: {}", mango_group.ref_surcharge_centibps, mango_group.ref_share_centibps, mango_group.ref_mngo_required);
+
+        // TODO - when this goes out, if there are any events on the EventQueue fee logging will be messed up
+
+        mango_group.ref_surcharge_centibps = ref_surcharge_centibps;
+        mango_group.ref_share_centibps = ref_share_centibps;
+        mango_group.ref_mngo_required = ref_mngo_required;
+
+        msg!("new referral fee params: ref_surcharge_centibps: {} ref_share_centibps: {} ref_mngo_required: {}", ref_surcharge_centibps, ref_share_centibps, ref_mngo_required);
+        Ok(())
+    }
+
+    #[inline(never)]
+    /// Store the referrer's MangoAccount pubkey on the Referrer account
+    /// It will create the Referrer account as a PDA of user's MangoAccount if it doesn't exist
+    /// This is primarily useful for the UI; the referrer address stored here is not necessarily
+    /// who earns the ref fees.
+    fn set_referrer_memory(program_id: &Pubkey, accounts: &[AccountInfo]) -> MangoResult {
+        const NUM_FIXED: usize = 7;
+        let [
+            mango_group_ai,             // read
+            mango_account_ai,           // read
+            owner_ai,                   // signer
+            referrer_memory_ai,         // write
+            referrer_mango_account_ai,  // read
+            payer_ai,                   // write, signer
+            system_prog_ai,             // read
+        ] = array_ref![accounts, 0, NUM_FIXED];
+        check!(
+            system_prog_ai.key == &solana_program::system_program::id(),
+            MangoErrorCode::InvalidProgramId
+        )?;
+
+        let _ = MangoGroup::load_checked(mango_group_ai, program_id)?;
+        let mango_account =
+            MangoAccount::load_checked(mango_account_ai, program_id, &mango_group_ai.key)?;
+        check!(
+            &mango_account.owner == owner_ai.key || &mango_account.delegate == owner_ai.key,
+            MangoErrorCode::InvalidOwner
+        )?;
+        check!(owner_ai.is_signer, MangoErrorCode::InvalidSignerKey)?;
+
+        let _ =
+            MangoAccount::load_checked(referrer_mango_account_ai, program_id, mango_group_ai.key)?;
+
+        if referrer_memory_ai.data_is_empty() {
+            // initialize it if it's not initialized yet
+            let referrer_seeds: &[&[u8]] = &[&mango_account_ai.key.as_ref(), b"ReferrerMemory"];
+            seed_and_create_pda(
+                program_id,
+                payer_ai,
+                &Rent::get()?,
+                size_of::<ReferrerMemory>(),
+                program_id,
+                system_prog_ai,
+                referrer_memory_ai,
+                referrer_seeds,
+                &[],
+            )?;
+            ReferrerMemory::init(referrer_memory_ai, program_id, referrer_mango_account_ai)
+        } else {
+            // otherwise just set referrer pubkey
+            let mut referrer_memory =
+                ReferrerMemory::load_mut_checked(referrer_memory_ai, program_id)?;
+            referrer_memory.referrer_mango_account = *referrer_mango_account_ai.key;
+            Ok(())
+        }
+    }
+
+    /// Associate the referrer's MangoAccount with a human readable `referrer_id` which can be used
+    /// in a ref link
+    /// Create the `ReferrerIdRecord` PDA; if it already exists throw error
+    #[inline(never)]
+    fn register_referrer_id(
+        program_id: &Pubkey,
+        accounts: &[AccountInfo],
+        referrer_id: [u8; INFO_LEN],
+    ) -> MangoResult {
+        const NUM_FIXED: usize = 5;
+        let [
+            mango_group_ai,             // read
+            referrer_mango_account_ai,  // read
+            referrer_id_record_ai,      // write
+            payer_ai,                   // write, signer
+            system_prog_ai,             // read
+        ] = array_ref![accounts, 0, NUM_FIXED];
+        check!(
+            system_prog_ai.key == &solana_program::system_program::id(),
+            MangoErrorCode::InvalidProgramId
+        )?;
+
+        let _ = MangoGroup::load_checked(mango_group_ai, program_id)?;
+
+        let _ =
+            MangoAccount::load_checked(referrer_mango_account_ai, program_id, mango_group_ai.key)?;
+
+        // referrer_id_record must be empty; cannot be transferred
+        check!(referrer_id_record_ai.data_is_empty(), MangoErrorCode::InvalidAccount)?;
+        let referrer_record_seeds: &[&[u8]] =
+            &[&mango_group_ai.key.as_ref(), b"ReferrerIdRecord", &referrer_id];
+        seed_and_create_pda(
+            program_id,
+            payer_ai,
+            &Rent::get()?,
+            size_of::<ReferrerIdRecord>(),
+            program_id,
+            system_prog_ai,
+            referrer_id_record_ai,
+            referrer_record_seeds,
+            &[],
+        )?;
+
+        ReferrerIdRecord::init(
+            referrer_id_record_ai,
+            program_id,
+            referrer_mango_account_ai,
+            referrer_id,
+        )
+    }
     pub fn process(program_id: &Pubkey, accounts: &[AccountInfo], data: &[u8]) -> MangoResult {
         let instruction =
             MangoInstruction::unpack(data).ok_or(ProgramError::InvalidInstructionData)?;
@@ -6141,6 +6279,28 @@ impl Processor {
                 msg!("Mango: CreateSpotOpenOrders");
                 Self::create_spot_open_orders(program_id, accounts)
             }
+            MangoInstruction::ChangeReferralFeeParams {
+                ref_surcharge_centibps,
+                ref_share_centibps,
+                ref_mngo_required,
+            } => {
+                msg!("Mango: ChangeReferralFeeParams");
+                Self::change_referral_fee_params(
+                    program_id,
+                    accounts,
+                    ref_surcharge_centibps,
+                    ref_share_centibps,
+                    ref_mngo_required,
+                )
+            }
+            MangoInstruction::SetReferrerMemory => {
+                msg!("Mango: SetReferrerMemory");
+                Self::set_referrer_memory(program_id, accounts)
+            }
+            MangoInstruction::RegisterReferrerId { referrer_id } => {
+                msg!("Mango: RegisterReferrerId");
+                Self::register_referrer_id(program_id, accounts, referrer_id)
+            }
         }
     }
 }
@@ -6309,9 +6469,11 @@ fn read_oracle(
             #[cfg(not(feature = "devnet"))]
             let conf = I80F48::from_num(price_account.agg.conf).checked_div(value).unwrap();
 
-            // Pyth status check temporarily removed to let people use accounts with COPE
             #[cfg(not(feature = "devnet"))]
-            if conf > PYTH_CONF_FILTER {
+            if price_account.agg.status != PriceStatus::Trading {
+                msg!("Pyth status invalid: {}", price_account.agg.status as u8);
+                return Err(throw_err!(MangoErrorCode::InvalidOraclePrice));
+            } else if conf > PYTH_CONF_FILTER {
                 msg!(
                     "Pyth conf interval too high; oracle index: {} value: {} conf: {}",
                     token_index,
@@ -6784,7 +6946,7 @@ fn seed_and_create_pda<'a>(
     pda_account: &AccountInfo<'a>,
     seeds: &[&[u8]],
     funder_seeds: &[&[u8]],
-) -> ProgramResult {
+) -> MangoResult {
     let (pda_address, bump) = Pubkey::find_program_address(seeds, program_id);
     check!(&pda_address == pda_account.key, MangoErrorCode::InvalidAccount)?;
     create_pda_account(
@@ -6796,7 +6958,8 @@ fn seed_and_create_pda<'a>(
         pda_account,
         &[seeds, &[&[bump]]].concat(),
         funder_seeds,
-    )
+    )?;
+    Ok(())
 }
 fn create_pda_account<'a>(
     funder: &AccountInfo<'a>,
@@ -6889,24 +7052,4 @@ pub fn get_leverage_weights(leverage: I80F48) -> (I80F48, I80F48) {
         (leverage - ONE_I80F48).checked_div(leverage).unwrap(),
         (leverage + ONE_I80F48).checked_div(leverage).unwrap(),
     )
-}
-
-pub fn emit_perp_balances(
-    mango_group: Pubkey,
-    mango_account: Pubkey,
-    market_index: u64,
-    pa: &PerpAccount,
-    perp_market_cache: &PerpMarketCache,
-) {
-    mango_emit!(PerpBalanceLog {
-        mango_group: mango_group,
-        mango_account: mango_account,
-        market_index: market_index,
-        base_position: pa.base_position,
-        quote_position: pa.quote_position.to_bits(),
-        long_settled_funding: pa.long_settled_funding.to_bits(),
-        short_settled_funding: pa.short_settled_funding.to_bits(),
-        long_funding: perp_market_cache.long_funding.to_bits(),
-        short_funding: perp_market_cache.long_funding.to_bits(),
-    });
 }