+++ program/src/state.rs
@@ -47,6 +47,7 @@ pub const FREE_ORDER_SLOT: u8 = u8::MAX;
 pub const MAX_NUM_IN_MARGIN_BASKET: u8 = 9;
 pub const INDEX_START: I80F48 = I80F48!(1_000_000);
 pub const PYTH_CONF_FILTER: I80F48 = I80F48!(0.10); // filter out pyth prices with conf > 10% of price
+pub const CENTIBPS_PER_UNIT: I80F48 = I80F48!(1_000_000);
 
 declare_check_assert_macros!(SourceFileId::State);
 
@@ -78,6 +79,8 @@ pub enum DataType {
     MangoCache,
     EventQueue,
     AdvancedOrders,
+    ReferrerMemory,
+    ReferrerIdRecord,
 }
 
 const NUM_HEALTHS: usize = 2;
@@ -204,7 +207,10 @@ pub struct MangoGroup {
     pub max_mango_accounts: u32, // limits maximum number of MangoAccounts.v1 (closeable) accounts
     pub num_mango_accounts: u32, // number of MangoAccounts.v1
 
-    pub padding: [u8; 24], // padding used for future expansions
+    pub ref_surcharge_centibps: u32, // 100
+    pub ref_share_centibps: u32,     // 80 (must be less than surcharge)
+    pub ref_mngo_required: u64,
+    pub padding: [u8; 8], // padding used for future expansions
 }
 
 impl MangoGroup {
@@ -247,6 +253,9 @@ impl MangoGroup {
     pub fn find_root_bank_index(&self, root_bank_pk: &Pubkey) -> Option<usize> {
         // TODO profile and optimize
         self.tokens.iter().position(|token_info| &token_info.root_bank == root_bank_pk)
+    }
+    pub fn find_token_index(&self, mint_pk: &Pubkey) -> Option<usize> {
+        self.tokens.iter().position(|token_info| &token_info.mint == mint_pk)
     }
     pub fn find_spot_market_index(&self, spot_market_pk: &Pubkey) -> Option<usize> {
         self.spot_markets
@@ -1061,8 +1070,9 @@ impl HealthCache {
         bids_quantity: i64,
         asks_quantity: i64,
     ) -> MangoResult<I80F48> {
+        let info = &mango_group.perp_markets[market_index];
         let (base, quote) = mango_account.perp_accounts[market_index].sim_get_val(
-            &mango_group.perp_markets[market_index],
+            info,
             &mango_cache.perp_market_cache[market_index],
             mango_cache.price_cache[market_index].price,
             taker_base,
@@ -1093,7 +1103,27 @@ impl HealthCache {
         };
 
         let h = self.health[health_type as usize].ok_or(throw!())?;
-        Ok(h + curr_perp_health - prev_perp_health)
+
+        // Apply taker fees; Assume no referrer
+        let taker_fees = if taker_quote != 0 {
+            let taker_quote_native =
+                I80F48::from_num(info.quote_lot_size.checked_mul(taker_quote.abs()).unwrap());
+            let mut market_fees = info.taker_fee * taker_quote_native;
+            if let Some(mngo_index) = mango_group.find_token_index(&mngo_token::id()) {
+                let mngo_cache = &mango_cache.root_bank_cache[mngo_index];
+                let mngo_deposits = mango_account.get_native_deposit(mngo_cache, mngo_index)?;
+                let ref_mngo_req = I80F48::from_num(mango_group.ref_mngo_required);
+                if mngo_deposits < ref_mngo_req {
+                    market_fees += (I80F48::from_num(mango_group.ref_surcharge_centibps)
+                        / CENTIBPS_PER_UNIT)
+                        * taker_quote_native;
+                }
+            }
+            market_fees
+        } else {
+            ZERO_I80F48
+        };
+        Ok(h + curr_perp_health - prev_perp_health - taker_fees)
     }
 
     /// Update perp val and then update the healths
@@ -1561,7 +1591,6 @@ impl MangoAccount {
         &mut self,
         market_index: usize,
         perp_market: &mut PerpMarket,
-        info: &PerpMarketInfo,
         cache: &PerpMarketCache,
         fill: &FillEvent,
     ) -> MangoResult<()> {
@@ -1571,9 +1600,10 @@ impl MangoAccount {
         pa.remove_taker_trade(base_change, quote_change);
         pa.change_base_position(perp_market, base_change);
         let quote = I80F48::from_num(perp_market.quote_lot_size * quote_change);
-        let fees = quote.abs() * info.taker_fee;
-        perp_market.fees_accrued += fees;
-        pa.quote_position += quote - fees;
+
+        // fees are assessed at time of trade; no need to assess fees here
+
+        pa.quote_position += quote;
         Ok(())
     }
 
@@ -1581,7 +1611,6 @@ impl MangoAccount {
         &mut self,
         market_index: usize,
         perp_market: &mut PerpMarket,
-        info: &PerpMarketInfo,
         cache: &PerpMarketCache,
         fill: &FillEvent,
     ) -> MangoResult<()> {
@@ -1592,7 +1621,7 @@ impl MangoAccount {
         let (base_change, quote_change) = fill.base_quote_change(side);
         pa.change_base_position(perp_market, base_change);
         let quote = I80F48::from_num(perp_market.quote_lot_size.checked_mul(quote_change).unwrap());
-        let fees = quote.abs() * info.maker_fee;
+        let fees = quote.abs() * fill.maker_fee;
         perp_market.fees_accrued += fees;
         pa.quote_position = pa.quote_position.checked_add(quote - fees).unwrap();
 
@@ -2491,5 +2520,75 @@ impl AdvancedOrders {
         )?;
         check!(&mango_account.advanced_orders_key == account.key, MangoErrorCode::InvalidAccount)?;
         Ok(state)
+    }
+}
+
+/// Store the referrer's mango account
+#[derive(Copy, Clone, Pod, Loadable)]
+#[repr(C)]
+pub struct ReferrerMemory {
+    pub meta_data: MetaData,
+    pub referrer_mango_account: Pubkey,
+}
+
+impl ReferrerMemory {
+    pub fn init(
+        account: &AccountInfo,
+        program_id: &Pubkey,
+        referrer_mango_account_ai: &AccountInfo,
+    ) -> MangoResult {
+        let mut state: RefMut<Self> = Self::load_mut(account)?;
+        check!(account.owner == program_id, MangoErrorCode::InvalidOwner)?;
+        check!(!state.meta_data.is_initialized, MangoErrorCode::InvalidAccountState)?;
+
+        state.meta_data = MetaData::new(DataType::ReferrerMemory, 0, true);
+        state.referrer_mango_account = *referrer_mango_account_ai.key;
+
+        Ok(())
+    }
+    pub fn load_mut_checked<'a>(
+        account: &'a AccountInfo,
+        program_id: &Pubkey,
+    ) -> MangoResult<RefMut<'a, Self>> {
+        // not really necessary because this is a PDA
+        check_eq!(account.owner, program_id, MangoErrorCode::InvalidOwner)?;
+
+        let state: RefMut<'a, Self> = Self::load_mut(account)?;
+
+        check!(state.meta_data.is_initialized, MangoErrorCode::InvalidAccountState)?;
+        check!(
+            state.meta_data.data_type == DataType::ReferrerMemory as u8,
+            MangoErrorCode::InvalidAccountState
+        )?;
+
+        Ok(state)
+    }
+}
+
+/// Register the referrer's id to be used in the URL
+#[derive(Copy, Clone, Pod, Loadable)]
+#[repr(C)]
+pub struct ReferrerIdRecord {
+    pub meta_data: MetaData,
+    pub referrer_mango_account: Pubkey,
+    pub id: [u8; INFO_LEN], // this id is one of the seeds
+}
+
+impl ReferrerIdRecord {
+    pub fn init(
+        account: &AccountInfo,
+        program_id: &Pubkey,
+        referrer_mango_account_ai: &AccountInfo,
+        referrer_id: [u8; INFO_LEN],
+    ) -> MangoResult {
+        let mut state: RefMut<Self> = Self::load_mut(account)?;
+        check!(account.owner == program_id, MangoErrorCode::InvalidOwner)?;
+        check!(!state.meta_data.is_initialized, MangoErrorCode::InvalidAccountState)?;
+
+        state.meta_data = MetaData::new(DataType::ReferrerIdRecord, 0, true);
+        state.referrer_mango_account = *referrer_mango_account_ai.key;
+        state.id = referrer_id;
+
+        Ok(())
     }
 }