+++ changelog.md
@@ -1,5 +1,11 @@
 # Mango Program Change Log
 
+## v3.3.4
+Deployed: | Slot:
+1. Added three instructions (ChangeReferralFeeParams, SetReferrerMemory, RegisterReferrerId) to help with referral program
+2. Assess taker fees at the time of the taker trade
+3. Add back Pyth status checks
+
 ## v3.3.3
 Deployed: Feb 4, 2022 at 01:47:33 UTC | Slot: 119,226,876
 1. Pyth status check temporarily removed to let people use accounts with COPE